Renaming 'exceptions' to 'exception'
